# docs/swagger.yaml
basePath: /v1
definitions:
...
  dtos.BasicUserInfo:
    properties:
      businesses_count:
        type: integer
      customers_count:
        type: integer
      loved_ones_count:
        type: integer
    type: object
...
  dtos.BusinessInfo:
    properties:
      customers_count:
        type: integer
      followers_count:
        type: integer
      is_followed:
        type: boolean
    type: object
...
  dtos.PagedRelationships:
    properties:
      limit:
        type: integer
      page:
        type: integer
      relationships:
        items:
          $ref: '#/definitions/dtos.UserRelationship'
        type: array
    type: object
...
  dtos.User:
    properties:
      country_code:
        type: string
      dob:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      interests:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
      relationship_status:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
...
  dtos.UserProfile:
    properties:
      account_type:
        type: integer
      basic_user_info:
        $ref: '#/definitions/dtos.BasicUserInfo'
      bio:
        type: string
      business_info:
        $ref: '#/definitions/dtos.BusinessInfo'
      celebrations_count:
        type: integer
      contents_count:
        type: integer
      country_code:
        type: string
      dob:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      interests:
        type: string
      is_verified:
        type: boolean
      last_name:
        type: string
      location:
        type: string
      phone_number:
        type: string
      profile_image:
        type: string
      relationship_status:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
  dtos.UserRelationship:
    type: object
  handlers.ErrorResponse:
    properties:
      developer_information:
        example: an error occurred
        type: string
      error:
        example: message
        type: string
      status:
        example: error
        type: string
    type: object
  handlers.editUserResponse:
    properties:
      status:
        type: string
      user:
        $ref: '#/definitions/dtos.User'
    type: object
  handlers.getUserResponse:
    properties:
      data:
        $ref: '#/definitions/dtos.UserProfile'
      status:
        type: string
    type: object
...
  relationships.followerResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  relationships.getFollowersResponse:
    properties:
      data:
        $ref: '#/definitions/dtos.PagedRelationships'
      status:
        type: string
    type: object
host: localhost:8083
info:
  contact: {}
  description: Celebut Backend REST endpoints
  title: Celebut API
  version: "1.0"
paths:
...
  /followers:
    get:
      consumes:
      - application/json
      description: Get businesses' followers
      operationId: get-followers
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/relationships.getFollowersResponse'
      security:
      - Bearer: []
      summary: Get followers
      tags:
      - User Relationships
  /followers/{userID}:
    post:
      consumes:
      - application/json
      description: Follow a business
      operationId: follow-business
      parameters:
      - description: user identifier
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/relationships.followerResponse'
      security:
      - Bearer: []
      summary: Follow business
      tags:
      - User Relationships
...
  /lovedones:
    get:
      consumes:
      - application/json
      description: Get user's loved ones
      operationId: get-loved-ones
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/relationships.getFollowersResponse'
      security:
      - Bearer: []
      summary: Get loved ones
      tags:
      - User Relationships
  /lovedones/{userID}:
    delete:
      consumes:
      - application/json
      description: Remove user's loved one
      operationId: remove-loved-one
      parameters:
      - description: user identifier
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/relationships.followerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - Bearer: []
      summary: Remove loved one
      tags:
      - User Relationships
    post:
      consumes:
      - application/json
      description: Become a loved one
      operationId: become-a-loved-one
      parameters:
      - description: user identifier
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/relationships.followerResponse'
      security:
      - Bearer: []
      summary: Become a loved one
      tags:
      - User Relationships
...
  /unfollow/{userID}:
    delete:
      consumes:
      - application/json
      description: Unfollow a business
      operationId: unfollow-business
      parameters:
      - description: user identifier
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/relationships.followerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - Bearer: []
      summary: Unfollow business
      tags:
      - User Relationships
  /user:
    put:
      consumes:
      - multipart/form-data
      description: Edit user's profile
      operationId: edit-user
      parameters:
      - description: profile image file
        in: formData
        name: profile_image
        type: file
      - example: "234"
        in: formData
        name: country_code
        type: string
      - example: "2020-10-12"
        in: formData
        name: dob
        required: true
        type: string
      - example: user@email.com
        in: formData
        name: email
        type: string
      - example: John
        in: formData
        name: first_name
        required: true
        type: string
      - example: Doe
        in: formData
        name: last_name
        required: true
        type: string
      - example: password
        in: formData
        name: password
        required: true
        type: string
      - example: "0712345678"
        in: formData
        name: phone_number
        type: string
      - example: johndoe
        in: formData
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.editUserResponse'
      security:
      - Bearer: []
      summary: Edit user info
      tags:
      - Users
  /users/{userID}:
    get:
      consumes:
      - application/json
      description: Get user profile info
      operationId: get-user-info
      parameters:
      - description: user identifier
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.getUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - Bearer: []
      summary: Get user
      tags:
      - Users
securityDefinitions:
  Auth-Token:
    description: Type token for unauthenticated requests
    in: header
    name: x-auth-token
    type: apiKey
  Bearer:
    description: '`Bearer token` for authenticated requests'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"